+++ src/main/java/finalmission/domain/TokenProcessor.java
@@ -1,0 +1,10 @@
+package finalmission.domain;
+
+import finalmission.dto.AuthenticatedMember;
+
+public interface TokenProcessor {
+
+    String createToken(TokenAuthRole tokenAuthRole, Long id);
+
+    AuthenticatedMember extract(String rawToken);
+}
