+++ src/main/java/finalmission/global/web/CheckUserInterceptor.java
@@ -2,8 +2,8 @@ package finalmission.global.web;
 
 import static finalmission.domain.TokenAuthRole.USER;
 
-import finalmission.domain.AuthenticatedMember;
-import finalmission.service.TokenService;
+import finalmission.dto.AuthenticatedMember;
+import finalmission.domain.TokenProcessor;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
@@ -14,13 +14,13 @@ import org.springframework.web.servlet.HandlerInterceptor;
 @RequiredArgsConstructor
 public class CheckUserInterceptor implements HandlerInterceptor {
 
-    private final TokenService tokenService;
+    private final TokenProcessor tokenProcessor;
 
     @Override
     public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler)
             throws Exception {
         String rawToken = request.getHeader("Authorization");
-        AuthenticatedMember authenticatedMember = tokenService.extract(rawToken);
+        AuthenticatedMember authenticatedMember = tokenProcessor.extract(rawToken);
         if (authenticatedMember.tokenAuthRole().hasAccessRole(USER)) {
             return true;
         }
