+++ src/main/java/finalmission/service/CrewService.java
@@ -3,6 +3,7 @@ package finalmission.service;
 import static finalmission.domain.TokenAuthRole.USER;
 
 import finalmission.domain.Crew;
+import finalmission.domain.TokenProcessor;
 import finalmission.dto.request.CrewLoginRequest;
 import finalmission.dto.request.CrewSignUpRequest;
 import finalmission.dto.response.CrewLoginResponse;
@@ -18,13 +19,13 @@ import org.springframework.transaction.annotation.Transactional;
 public class CrewService {
 
     private final CrewRepository crewRepository;
-    private final TokenService tokenService;
+    private final TokenProcessor tokenProcessor;
     private final MeetingRepository meetingRepository;
 
     public CrewLoginResponse login(CrewLoginRequest request) {
         Crew crew = crewRepository.findByEmailAndPassword(request.email(), request.password())
                 .orElseThrow(() -> new IllegalArgumentException("크루의 이메일 혹은 비밀번호가 틀렸습니다."));
-        String token = tokenService.createToken(USER, crew.getId());
+        String token = tokenProcessor.createToken(USER, crew.getId());
         return new CrewLoginResponse(token);
     }
