--- src/main/java/finalmission/service/MailService.java
@@ -1,8 +1,0 @@
-package finalmission.service;
-
-public interface MailService {
-
-    void sendAcceptEmail(String acceptedEmail);
-
-    void sendDenyEmail(String deniedEmail);
-}
