--- src/main/java/finalmission/service/TokenService.java
@@ -1,11 +1,0 @@
-package finalmission.service;
-
-import finalmission.domain.AuthenticatedMember;
-import finalmission.domain.TokenAuthRole;
-
-public interface TokenService {
-
-    String createToken(TokenAuthRole tokenAuthRole, Long id);
-
-    AuthenticatedMember extract(String rawToken);
-}
